Add Policy to Parent Income Module

+++ app/Http/Controllers/API/ParentIncomeController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\ParentIncome\StoreParentIncomeRequest;
 use App\Http\Requests\ParentIncome\UpdateParentIncomeRequest;
 use App\Libs\Response\ResponseJSON;
+use App\Models\ParentIncome;
 use App\Repositories\EloquentParentIncomeRepository;
 use Illuminate\Http\Request;
 
@@ -16,6 +17,8 @@ class ParentIncomeController extends Controller
     public function __construct(EloquentParentIncomeRepository $parentIncomeRepo)
     {
         $this->parentIncomeRepo = $parentIncomeRepo;
+
+        $this->middleware('auth:sanctum');
     }
     /**
      * Display a listing of the resource.
@@ -24,6 +27,8 @@ class ParentIncomeController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', ParentIncome::class);
+
         $incomes = $this->parentIncomeRepo->getParentIncomes();
 
         return ResponseJSON::successWithData('Parent Incomes has been loaded', $incomes);
@@ -37,6 +42,8 @@ class ParentIncomeController extends Controller
      */
     public function store(StoreParentIncomeRequest $request)
     {
+        $this->authorize('create', ParentIncome::class);
+
         $requests = $request->validated();
 
         $res = $this->parentIncomeRepo->storeParentIncome($requests);
@@ -56,6 +63,8 @@ class ParentIncomeController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view', ParentIncome::find($id));
+
         $income = $this->parentIncomeRepo->getParentIncome($id);
 
         return ResponseJSON::successWithData('Parent Income has been loaded', $income);
@@ -70,6 +79,8 @@ class ParentIncomeController extends Controller
      */
     public function update(UpdateParentIncomeRequest $request, $id)
     {
+        $this->authorize('update', ParentIncome::find($id));
+
         $requests = $request->validated();
 
         $res = $this->parentIncomeRepo->updateParentIncome($requests, $id);
@@ -89,6 +100,8 @@ class ParentIncomeController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', ParentIncome::find($id));
+
         $res = $this->parentIncomeRepo->destroyParentIncome($id);
 
         if (!$res) {
